migration to python-ldap - correction on tree deletion (#28)
