+++ requirements-dev.txt
@@ -6,8 +6,8 @@ PyGithub
 GitPython
 pyjwt
 python-dotenv
+python-ldap
 Jinja2
-git+https://github.com/abilian/ldif.git@4.2.0
 pre-commit
 black
 flake8
