+++ requirements.txt
@@ -1,10 +1,12 @@
 click==8.1.6
-ldap3
+ldap3~=2.9.1
 oracledb==1.4
 ansible-runner
 PyGithub
 GitPython==3.1.37
-pyjwt
-python-dotenv
-Jinja2
-git+https://github.com/abilian/ldif.git@4.2.0
+pyjwt~=2.8.0
+python-dotenv~=1.0.0
+Jinja2~=3.1.2
+python-ldap
+requests~=2.31.0
+setuptools~=68.2.2
